routes test

+++ src/unbuild.py
@@ -3,4 +3,5 @@ import os
 
 if __name__ == "__main__":
     for template_name in loader.list_templates():
-        os.remove(f'{build_dir}/{template_name}')
+        if os.path.exists(f'{build_dir}/{template_name}'):
+            os.remove(f'{build_dir}/{template_name}')
